fix(SPM-2055): Select all on disabled advisories
Steps to Reproduce:
Navigate to the Advosry details page
Apply status filter-select Applicable
then select all() from filter

This PR correctly avoids selecting systems with the status, "Applicable", which should be disabled. Also check this functionality by going to Inventory -> Systems -> choose a system -> Click Patch tab. This should show the same selection functionality.

This PR also fixes a bug not mentioned in the Jira card. There is a bug on Patch -> Systems where if you select a system and then you go to any page after page 1 and `Select all`, it will only select advisories on that page and after. It will not select any items on previous pages.