fix(RHINENG-391): Create func for various newline scenarios
To see what's broken check the 2 following advisories:
RHBA-2023:2052
RHSA-2023:3246

We have 2 different ways that descriptions come in for Advisories. The most common way is to separate every line with `\n\n`. This would create double space between each line. To remedy this, we have a `preserveNewlines` function in `Helpers.js` that removes the extra `\n`.

But, the new variation of description correctly provides `\n\n` for paragraph separations and only `\n` for bullet separations. But with the current `preserveNewlines` functionality, it removes the `\n` for the bullets, making them appear as a single line (see RHBA-2023:2052).

This PR adds an additional function that takes the substring of the description that includes the bullets and replaces any `\n\n` with a single `\n`, ignoring the final `\n\n` because it is for the paragraph that comes after the bullets.

I expect this to fix most, if not all, of our current descriptions. If a new layout for descriptions comes through at some point, it could be broken. The best fix is to standardize the descriptions for all advisories.